+++ lueur/platform/gcp/cloudrun.py
@@ -184,6 +184,37 @@ def expand_links(d: Discovery, serialized: dict[str, Any]) -> None:
                     ),
                 )
 
+        nics = tpl.get("vpcAccess", {}).get("networkInterfaces")
+        if nics:
+            for nic in nics:
+                network = nic["network"]
+                p = f"$.resources[?@.meta.kind=='network' && @.meta.name=='{network}']"  # noqa E501
+                for nk in iter_resource(serialized, p):
+                    add_link(
+                        d,
+                        r_id,
+                        Link(
+                            direction="out",
+                            kind="network",
+                            path=nk.path,
+                            pointer=str(nk.pointer()),
+                        ),
+                    )
+
+                subnet = nic["subnetwork"]
+                p = f"$.resources[?@.meta.kind=='subnet' && @.meta.name=='{subnet}']"  # noqa E501
+                for sn in iter_resource(serialized, p):
+                    add_link(
+                        d,
+                        r_id,
+                        Link(
+                            direction="out",
+                            kind="subnet",
+                            path=sn.path,
+                            pointer=str(sn.pointer()),
+                        ),
+                    )
+
         for v in tpl.get("volumes", []):
             if v.get("name") != "cloudsql":
                 continue